Refactoring work on NBT
