+++ mcidle/src/networking/types/nbt/TagPod.cpp
@@ -1,0 +1,7 @@
+#include <networking/types/nbt/TagPod.hpp>
+
+namespace mcidle {
+namespace nbt {
+
+}
+}
